requirements.txt
torch>=2.0.0
transformers>=4.50.0
numpy>=1.24.0
nltk>=3.8.0
pytesseract>=0.3.10
pdf2image>=1.16.3
fpdf>=1.7.2
scikit-learn>=1.2.2
python-docx>=0.8.11
PyPDF2>=3.1.0
pyspellchecker>=0.7.5
opencv-python>=4.8.1.78
Pillow>=9.5.0
requests>=2.31.0
Flask>=2.2.0
Werkzeug>=2.2.0
